FB-15a: split attemptApiRequest into TaskApiRequestAttempt.ts
